+++ src/update.js
@@ -8,7 +8,7 @@ import {formatPercent, fx, fy, fh, fw, collapsePath} from './utilities';
 import {interpolatePath} from 'd3-interpolate-path';
 import {getTransitions} from './transitions';
 import {default as glb} from './globals';
-import {textColor, fill} from './color';
+import {textColor, fill, subBarColors} from './color';
 import {getFont} from './font';
 
 /**
@@ -28,9 +28,9 @@ export function update (bars, noDelay = false) {
     subBars.enter()
         .append('g')
             .attr('class', 'biPartite-subBar')
-            .attr('transform', d => {return `translate(${d.x}, ${d.y})`})
+            .attr('transform', d => `translate(${d.x}, ${d.y})`)
             .append('rect')
-                .style('fill', d => {return fill(d)})
+                .style('fill', d => fill(d))
                 .style('shape-rendering', 'crispEdges')
                 .transition(t)
                     .attr('x', fx)
@@ -46,9 +46,9 @@ export function update (bars, noDelay = false) {
 
     subBars
         .transition(t)
-        .attr('transform', d => {return `translate(${d.x}, ${d.y})`})
+        .attr('transform', d => `translate(${d.x}, ${d.y})`)
         .select('rect')
-            .style('fill', d => {return fill(d)})
+            .style('fill', d => fill(d))
             .attr('x', fx)
             .attr('y', fy)
             .attr('width', fw)
@@ -62,13 +62,10 @@ export function update (bars, noDelay = false) {
         .append('path')
         // .merge(edges)
         .attr('class','biPartite-edge')
-        .style('fill', d => {return fill(d)})
+        .style('fill', d => fill(d))
         .style('fill-opacity', glb.graph.edgeOpacity())
-        .attr('d', d => {return collapsePath(d.path)})
+        .attr('d', d => collapsePath(d.path))
         .transition(t)
-            // .attrTween('d', function (d) {
-            //     return interpolatePath(d3.select(this).attr('d'), d.path);                
-            // });
             .attrTween('d', (d, i, nodes) => {
                 return interpolatePath(d3.select(nodes[i]).attr('d'), d.path);                
             });
@@ -81,7 +78,7 @@ export function update (bars, noDelay = false) {
 
     edges
         .transition(t)
-            .style('fill', d => {return fill(d)})
+            .style('fill', d => fill(d))
             .attrTween('d', (d, i, nodes) => {
                 return interpolatePath(d3.select(nodes[i]).attr('d'), d.path);                
             });
@@ -92,14 +89,14 @@ export function update (bars, noDelay = false) {
        
     let mainBarsEnter = mainBars.enter()
         .append('g')
-            .attr('transform', d => {return `translate(${d.x}, ${d.y})`})
+            .attr('transform', d => `translate(${d.x}, ${d.y})`)
             .attr('class','biPartite-mainBar')
             .on(glb.eventTypeOver, glb.eventListenerOver)
             .on(glb.eventTypeOut, glb.eventListenerOut);
 
     mainBarsEnter
         .append('rect')
-            .attr('id', (d, i) => {return `mainBar${i}`})
+            .attr('id', (d, i) => `mainBar${i}`)
             .attr('x', fx)
             .attr('y', fy)
             .attr('width', fw)
@@ -112,11 +109,33 @@ export function update (bars, noDelay = false) {
             .attr('x', d => (d.part == 'primary'? -glb.LabelOffset : glb.LabelOffset))
             .attr('dy', '0.35em')
             .style('opacity', 0)
+            .style('fill', glb.labelColor)
+            .style('user-select', 'none')
             .text(d => d.key)
                 .attr('text-anchor', d => (d.part == 'primary' ? 'end': 'start'))
-                .attr('transform', d => {return glb.graph.orient() == 'vertical' ? 'rotate(0)' : 'rotate(90)'})
+                .attr('transform', d => glb.graph.orient() == 'vertical' ? 'rotate(0)' : 'rotate(90)')
                 .transition(t)
                     .style('opacity', 1);
+
+    mainBarsEnter.each ((d, i) => {
+            d3.select('#percentBlackG')
+                .append('g')
+                    .attr('id', `percentBlack${i}`)
+                    .attr('transform', `translate(${d.x}, ${d.y})`)
+                    .append('text')
+                        .attr('class', 'biPartite-percentage black')
+                        .attr('dy', '0.35em')
+                        .style('fill', glb.labelColor)
+                        .style('opacity', 0)
+                        .style('pointer-events', 'none')
+                        .style('user-select', 'none')
+                        .style('font', getFont('biPartite-percentage'))
+                        .text(formatPercent(d.percent))
+                            .attr('transform', glb.graph.orient() == 'vertical' ? 'rotate(0)' : 'rotate(90)')
+                            .attr('text-anchor', 'middle')
+                            .transition(t)
+                                .style('opacity', 1);
+    });
 
     mainBarsEnter
         .append('text')
@@ -124,15 +143,31 @@ export function update (bars, noDelay = false) {
             .attr('dy', '0.35em')
             .style('font', getFont('biPartite-percentage'))
             .style('opacity', 0)
+            .style('fill', d => {
+                if (d.part == 'primary') {
+                    return textColor([fill(d)]);            
+                } else {
+                    return textColor(subBarColors(bars.subBars, d));
+                };
+            })
             .style('pointer-events', 'none')
-            .style('clip-path', (d, i) => {return `url(#clip${i})`})
+            .style('user-select', 'none')
+            .style('clip-path', (d, i) => `url(#clip${i})`)
             .text(function(d){ return formatPercent(d.percent)})
-                .attr('transform', d => {return glb.graph.orient() == 'vertical' ? 'rotate(0)' : 'rotate(90)'})
+                .attr('transform', d => glb.graph.orient() == 'vertical' ? 'rotate(0)' : 'rotate(90)')
+                .attr('text-anchor', 'middle')
                 .transition(t)
                     .style('opacity', 1);
 
+    mainBars.each ((d, i) => {
+            d3.select(`#percentBlack${i}`)
+                .attr('transform', `translate(${d.x}, ${d.y})`)
+                .select('text')
+                    .text(formatPercent(d.percent));
+    });
+
     mainBars
-        .attr('transform', d => {return `translate(${d.x}, ${d.y})`})
+        .attr('transform', d => `translate(${d.x}, ${d.y})`)
         .select('rect')
             .attr('x', fx)
             .attr('y', fy)
@@ -144,19 +179,35 @@ export function update (bars, noDelay = false) {
             .transition(t1)
                 .style('opacity', 0)
             .transition(t2)
-                .attr('x', d => (d.part == 'primary'? -glb.LabelOffset : glb.LabelOffset))
+                .attr('x', d => d.part == 'primary'? -glb.LabelOffset : glb.LabelOffset)
                 .attr('text-anchor', d => (d.part == 'primary' ? 'end': 'start'))
                 .text(d => d.key)
                 .style('opacity', 1);
 
     mainBars
         .select('.biPartite-percentage.white')
-            .text(function(d) {return formatPercent(d.percent)})
+            .text(d => formatPercent(d.percent));
+
+    mainBarsEnter.each ((d, i) => {
+            d3.select('#svg').select('defs')
+                .append('clipPath')
+                    .attr('id', `clip${i}`)
+                    .append('use')
+                        .attr('xlink:href', `#mainBar${i}`);
+    });
 
     mainBars.exit()
         .transition(t)
             .on('start', (d, i, nodes) => {
                 d3.select(nodes[i]).remove()
             });
+
+    mainBars.exit().each ((d, i) => {
+            d3.select(`#clip${i}`)
+                .remove();
+            d3.select(`#percentBlack${i}`)
+                .remove();
+    });
+
  }