+++ test/src/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-Object.defineProperty(exports, '__esModule', { value: true });
-
-var biPartite = require('./biPartite.js');
-
-
-
-exports.biPartite = biPartite.biPartite;